+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+* [FEATURE] read initialization params first from the location hash then fallback to the location query string
+
 ## v1.0.0-alpha.7 - 2017-06-14
 
 * [FIX] event matching fails if any event props pattern is not specified