+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Upcoming
 
 * [FEATURE] read initialization params first from the location hash then fallback to the location query string
+* [FEATURE] add Fetch API support for deskpro api request
 
 ## v1.0.0-alpha.7 - 2017-06-14